Fix Compose lint compliance - add modifier parameters to NutritionScreen composables

+++ app/src/main/java/com/example/fitapp/data/db/Dao.kt
@@ -382,25 +382,25 @@ interface MealEntryDao {
             END
         ) FROM meal_entries WHERE date = :date
     """)
-    suspend fun getTotalCaloriesForDate(date: String): Double?
+    suspend fun getTotalCaloriesForDate(date: String): Float?
 
     @Query("""
         SELECT SUM((quantityGrams / 100.0) * (SELECT carbs FROM food_items WHERE id = foodItemId))
         FROM meal_entries WHERE date = :date
     """)
-    suspend fun getTotalCarbsForDate(date: String): Double?
+    suspend fun getTotalCarbsForDate(date: String): Float?
 
     @Query("""
         SELECT SUM((quantityGrams / 100.0) * (SELECT protein FROM food_items WHERE id = foodItemId))
         FROM meal_entries WHERE date = :date
     """)
-    suspend fun getTotalProteinForDate(date: String): Double?
+    suspend fun getTotalProteinForDate(date: String): Float?
 
     @Query("""
         SELECT SUM((quantityGrams / 100.0) * (SELECT fat FROM food_items WHERE id = foodItemId))
         FROM meal_entries WHERE date = :date
     """)
-    suspend fun getTotalFatForDate(date: String): Double?
+    suspend fun getTotalFatForDate(date: String): Float?
 }
 
 @Dao
